requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
dj-database-url>=2.2,<2.4
gunicorn>=22.0,<24.0
psycopg2-binary>=2.9,<3.0
whitenoise>=6.7,<6.9
